feat(protocol-parser): add new schema helpers
