+++ packages/protocol-parser/src/exports/index.ts
@@ -3,3 +3,7 @@
  *
  * Be sure we're ready to commit to these being supported and changes made backward compatible!
  */
+
+export { getKeySchema } from "../getKeySchema";
+export { getValueSchema } from "../getValueSchema";
+export { getSchemaTypes } from "../getSchemaTypes";
