+++ packages/store/ts/config/v2/compat.ts
@@ -1,7 +1,7 @@
 import { conform } from "@arktype/util";
 import { Schema, Store, Table } from "./output";
 import { mapObject } from "@latticexyz/common/utils";
-import { getKeySchema, getValueSchema } from "./table";
+import { getKeySchema, getValueSchema } from "@latticexyz/protocol-parser";
 
 export type storeToV1<store> = store extends Store
   ? {
